refactor: split logic into hooks and fix minor bugs
