+++ src/components/Calendar/DatesGrid/DatesGrid.styled.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const StyledDatesList = styled.ul`
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+`;
