+++ src/components/Calendar/DatesGrid/index.tsx
@@ -1,0 +1,3 @@
+import { DatesGrid } from "./DatesGrid";
+
+export default DatesGrid;
