--- src/components/Calendar/WeekDay/WeekDay.stories.tsx
@@ -1,19 +1,0 @@
-import { WeekDay as Component } from "./WeekDay";
-
-import type { Meta, StoryObj } from "@storybook/react";
-
-const meta: Meta<typeof Component> = {
-  component: Component,
-  title: "WeekDay",
-  tags: ["autodocs"],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof Component>;
-
-export const Default: Story = {
-  args: {
-    weekDay: "Mo",
-  },
-};
