+++ src/components/Calendar/WeekDays/WeekDays.stories.tsx
@@ -1,10 +1,12 @@
-import { Month as Component } from "./Month";
+import { WEEK_DAYS_FROM_MONDAY } from "@constants/calendar";
+
+import { WeekDays as Component } from "./WeekDays";
 
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof Component> = {
   component: Component,
-  title: "Month",
+  title: "WeekDays",
   tags: ["autodocs"],
 };
 
@@ -14,6 +16,6 @@ type Story = StoryObj<typeof Component>;
 
 export const Default: Story = {
   args: {
-    month: "January",
+    weekDays: WEEK_DAYS_FROM_MONDAY,
   },
 };
