+++ src/components/Calendar/WeekDays/index.tsx
@@ -1,0 +1,3 @@
+import { WeekDays } from "./WeekDays";
+
+export default WeekDays;
