+++ src/components/Calendar/_hooks/useCalendarVisibility.ts
@@ -1,0 +1,14 @@
+import { useState } from "react";
+
+export const useCalendarVisibility = () => {
+  const [showCalendar, setShowCalendar] = useState(false);
+
+  const toggleShowCalendar = () => {
+    setShowCalendar(!showCalendar);
+  };
+
+  return {
+    showCalendar,
+    toggleShowCalendar,
+  };
+};
