+++ src/components/DateRangePicker/DateRangePicker.tsx
@@ -1,15 +1,21 @@
 import { FC } from "react";
+import { DefaultTheme } from "styled-components/dist/types";
 
 import Calendar from "@components/Calendar";
 import ThemeWrapper from "@components/ThemeWrapper";
 import ErrorBoundary from "@components/ErrorBoundary";
-import { CalendarProps } from "@components/Calendar/Calendar";
+import { CalendarProps } from "@components/Calendar/Calendar.types";
 
-interface DateRangePickerProps extends CalendarProps {}
+interface DateRangePickerProps extends CalendarProps {
+  theme?: DefaultTheme;
+}
 
-export const DateRangePicker: FC<DateRangePickerProps> = (props) => (
+export const DateRangePicker: FC<DateRangePickerProps> = ({
+  theme,
+  ...props
+}) => (
   <ErrorBoundary>
-    <ThemeWrapper>
+    <ThemeWrapper theme={theme}>
       <Calendar {...props} withRangePicker={true} />
     </ThemeWrapper>
   </ErrorBoundary>
