+++ src/components/ThemeWrapper/index.tsx
@@ -1,13 +1,20 @@
 import { FC, PropsWithChildren } from "react";
-import { ThemeProvider } from "styled-components";
+import { DefaultTheme, ThemeProvider } from "styled-components";
 
 import { GlobalStyles } from "@styles/globalStyles";
 import { lightTheme } from "@styles/theme";
 
 import "@styles/Font.css";
 
-const ThemeWrapper: FC<PropsWithChildren> = ({ children }) => (
-  <ThemeProvider theme={lightTheme}>
+interface ThemeWrapperProps extends PropsWithChildren {
+  theme?: DefaultTheme;
+}
+
+const ThemeWrapper: FC<ThemeWrapperProps> = ({
+  children,
+  theme = lightTheme,
+}) => (
+  <ThemeProvider theme={theme}>
     <GlobalStyles />
     {children}
   </ThemeProvider>
