+++ src/components/TodoList/TodoList.tsx
@@ -3,7 +3,7 @@ import { ChangeEvent, FC, useState } from "react";
 import IconButton from "@components/IconButton";
 import Clear from "@components/Icons/Clear";
 
-import { useTodo } from "../../decorators/withToDo/TodoProvider";
+import { useTodo } from "@decorators/withToDo/TodoProvider";
 
 import TodoItem from "./TodoItem";
 import {
