+++ src/constants/constants.ts
@@ -16,3 +16,7 @@ export const INPUT_REGEX = /^(0[1-9]|[12][0-9]|3[01])\/(0[1-9]|1[0-2])\/\d{4}$/;
 export const Z_INDEX_MANAGER = {
   TODOLIST: 1000,
 };
+
+export const TODAY = new Date();
+
+export const EMPTY_CALLBACK = () => {};
