--- src/declarations.d.ts
@@ -1,5 +1,0 @@
-declare module "*.svg" {
-  const content: string;
-  export const ReactComponent;
-  export default content;
-}
