+++ src/decorators/withClearButton/index.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 
-import { CalendarProps } from "@components/Calendar/Calendar";
+import { CalendarProps } from "@components/Calendar/Calendar.types";
 
 const withClearButton = (Component: FC<CalendarProps>): FC<CalendarProps> => {
   const Wrappee = (props: CalendarProps) => {
