+++ src/styles/Font.css
@@ -1,4 +1,4 @@
 /* Font */
-/* Styled components has bad support for @import statement so we neeed to import it in a css file */
 
+/* Styled components have bad support for @import statement so we neeed to import it in a css file */
 @import url("https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
