+++ src/types/index.d.ts
@@ -1,3 +1,9 @@
+declare module "*.svg" {
+  const content: string;
+  export const ReactComponent;
+  export default content;
+}
+
 type Day = number;
 type WeekDay = string;
 type MonthWord = string;
