+++ vercel.json
@@ -6,3 +6,4 @@
   "framework": null,
   "outputDirectory": "./storybook-static"
 }
+
